all logic to show Gab question and answer moved to Gab Component
