+++ src/components/Gab/Gab.scss
@@ -1,0 +1,3 @@
+.gab {
+  color: red;
+}
